Update inspectmodules.py
Added Variables to the tree.
At the TextCtrl entry, hit ENTER to get to data through the tree.
In the tree, when you are at an object's item, hit SPACEBAR to get more info through outpout TextCtrl. On only one named treectrl button, Modules, hit SPACEBAR on that name and it go to output with list of modules linked with the module you are inspecting.